refactor(utils): break out utility functions
