+++ src/utils/createReducer.js
@@ -1,0 +1,10 @@
+export function createReducer (initialState, fnMap) {
+  return (state = initialState, action) => {
+    const { type, payload } = action;
+    const handler = fnMap[type];
+
+    return handler ? handler(state, payload) : state;
+  };
+}
+
+export default createReducer;
